requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
daphne==4.0.0
Django==4.0.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
fake-useragent==0.1.11
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==22.10.0
lxml==4.8.0
parse==1.19.0
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyppeteer==1.0.2
pyquery==1.4.3
pytz==2022.1
requests==2.27.1
requests-html==0.10.0
service-identity==21.1.0
six==1.16.0
slugify==0.0.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.4.0
tzdata==2022.1
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
zipp==3.8.0
zope.interface==5.5.0
